added confirmation for deleting company

+++ client/src/app/company/company.component.ts
@@ -1,35 +1,34 @@
-
-import { Component, Input, OnInit, Output } from '@angular/core';
-import { Company } from '../_models/Company'
-import { CreateCompanyComponent } from '../create-company/create-company.component'
+import { Component, Input, OnInit, Output } from "@angular/core";
+import { Company } from "../_models/Company";
+import { CreateCompanyComponent } from "../create-company/create-company.component";
 import { MatDialog, MatDialogConfig } from "@angular/material";
-import { CompanyService } from '../_services/company.service';
-import { NotificationService } from '../_services/notification.service';
-import { ManageCompaniesComponent } from '../manage-companies/manage-companies.component';
-
+import { CompanyService } from "../_services/company.service";
+import { NotificationService } from "../_services/notification.service";
+import { ManageCompaniesComponent } from "../manage-companies/manage-companies.component";
+import { ConfirmDialogComponent } from "../confirm-dialog/confirm-dialog.component";
 
 @Component({
-  selector: 'company-component',
-  templateUrl: './company.component.html',
-  styleUrls: ['./company.component.css']
+  selector: "company-component",
+  templateUrl: "./company.component.html",
+  styleUrls: ["./company.component.css"],
 })
 export class CompanyComponent implements OnInit {
   @Input() company: Company;
 
-  mailTo:string;
-  
+  mailTo: string;
+
   public displaySendEmailState: Boolean = false;
 
-  constructor(private companyService: CompanyService,
+  constructor(
+    private companyService: CompanyService,
     private notificationService: NotificationService,
-    private manageCompanies: ManageCompaniesComponent
-  ) { }
+    private manageCompanies: ManageCompaniesComponent,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit() {
-
     this.mailTo = this.getMailTo();
   }
-
 
   getMailTo(): string {
     const { username, password } = this.company.companyUser;
@@ -38,25 +37,44 @@ export class CompanyComponent implements OnInit {
   }
 
   copyToClip() {
-
-    const selBox = document.createElement('textarea');
-    selBox.style.position = 'fixed';
-    selBox.style.left = '0';
-    selBox.style.top = '0';
-    selBox.style.opacity = '0';
+    const selBox = document.createElement("textarea");
+    selBox.style.position = "fixed";
+    selBox.style.left = "0";
+    selBox.style.top = "0";
+    selBox.style.opacity = "0";
     selBox.value = `Username: ${this.company.companyUser.username}\nPassword: ${this.company.companyUser.password}`;
     document.body.appendChild(selBox);
     selBox.focus();
     selBox.select();
-    document.execCommand('copy');
+    document.execCommand("copy");
     document.body.removeChild(selBox);
   }
 
   delete() {
-    this.companyService.deleteCompany(this.company).subscribe((result) => {
-      this.notificationService.showNotif(this.company.name + ' is deleted', 'DELETE');
-      this.manageCompanies.loadCompanies();
-    })
+    this.openDialog();
+  }
+
+  openDialog() {
+    const dialogConfig = new MatDialogConfig();
+
+    dialogConfig.data = {
+      companyName: this.company.name,
+    };
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, dialogConfig);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.companyService.deleteCompany(this.company).subscribe((result) => {
+          this.notificationService.showNotif(
+            this.company.name + " is deleted",
+            "DELETE"
+          );
+          this.manageCompanies.loadCompanies();
+        });
+      } else {
+      }
+    });
   }
 
   displaySendEmail(): void {
